refactor: create initial character_info layout
