+++ app/src/main/java/io/github/rosariopfernandes/minibrothereye/ui/characterinfo/CharacterInfoFragment.kt
@@ -1,4 +1,4 @@
-package io.github.rosariopfernandes.minibrothereye.ui
+package io.github.rosariopfernandes.minibrothereye.ui.characterinfo
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -12,21 +12,18 @@ import io.github.rosariopfernandes.minibrothereye.R
 /**
  * A simple [Fragment] subclass as the second destination in the navigation.
  */
-class SecondFragment : Fragment() {
+class CharacterInfoFragment : Fragment() {
 
     override fun onCreateView(
             inflater: LayoutInflater, container: ViewGroup?,
             savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_second, container, false)
+        return inflater.inflate(R.layout.fragment_character_info, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        view.findViewById<Button>(R.id.button_second).setOnClickListener {
-            findNavController().navigate(R.id.action_SecondFragment_to_FirstFragment)
-        }
     }
 }
