+++ app/src/main/res/values/strings.xml
@@ -10,4 +10,10 @@
 
     <string name="hello_second_fragment">Hello second fragment. Arg: %1$s</string>
     <string name="error_cant_load_data">Oops! We couldnt load the data :(</string>
+
+    <string name="label_gender">Gender</string>
+    <string name="label_height">Height</string>
+    <string name="label_weight">Weight</string>
+
+    <string name="label_biography">Biography</string>
 </resources>
